+++ src/libraries/BCAL/DBCALClump_factory.h
@@ -43,7 +43,7 @@ class DBCALClump_factory : public JFactory<DBCALClump> {
   
   jerror_t init(void);
   jerror_t brun(JEventLoop *loop, int32_t runnumber);
-  jerror_t evnt(JEventLoop *loop, int eventnumber);
+  jerror_t evnt(JEventLoop *loop, uint64_t eventnumber);
   
 };
