+++ src/libraries/PAIR_SPECTROMETER/DPSCHit_factory.cc
@@ -45,7 +45,7 @@ jerror_t DPSCHit_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DPSCHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
+jerror_t DPSCHit_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber)
 {
   // Only print messages for one thread whenever run number change
   static pthread_mutex_t print_mutex = PTHREAD_MUTEX_INITIALIZER;
@@ -129,7 +129,7 @@ jerror_t DPSCHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t DPSCHit_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t DPSCHit_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
   /// Generate DPSCHit object for each DPSCDigiHit object.
   /// This is where the first set of calibration constants
