+++ src/libraries/TRACKING/DTrackWireBased_factory.cc
@@ -86,7 +86,7 @@ jerror_t DTrackWireBased_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DTrackWireBased_factory::brun(jana::JEventLoop *loop, int runnumber)
+jerror_t DTrackWireBased_factory::brun(jana::JEventLoop *loop, int32_t runnumber)
 {
   // Get the geometry
   DApplication* dapp=dynamic_cast<DApplication*>(loop->GetJApplication());
@@ -164,7 +164,7 @@ jerror_t DTrackWireBased_factory::brun(jana::JEventLoop *loop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t DTrackWireBased_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t DTrackWireBased_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
   if(!fitter)return NOERROR;
